Crate oak_restricted_kernel_dice

This crate contains the logic used by oak_restricted_kernel to create attestations. It is broken out into a separate crate to allow this logic to be used independently used in tests to create mock attestations, without also pulling in oak_restricted_kernel’s allocator, which breaks tests.

Functions

generate_derived_key

generate_dice_data

Generates attestation evidence for the ‘measurement’ of the application.

measure_digest_sha2_256

Type Aliases

DerivedKey

A derived sealing key.

DigestSha2_256